help with downloading items
First I created a new directory to make a destination for the items we will download. Then I used the create function and then the download function that I found on the api. I am not sure if I crated the instance or not. And I am uncertain as what to put in place of "asset_types".